#1992 - Moved components to components folder.
